move template to another file
