+++ internal/extensions/actions/common/configurator.go
@@ -1,11 +1,6 @@
 package common
 
 import (
-	"bytes"
-	"fmt"
-	"strings"
-	"text/template"
-
 	"github.com/kyma-project/cli.v3/internal/clierror"
 	"github.com/kyma-project/cli.v3/internal/extensions/types"
 	"gopkg.in/yaml.v3"
@@ -17,40 +12,21 @@ type TemplateConfigurator[T any] struct {
 	Cfg T
 }
 
-var funcMap = template.FuncMap{
-	"newLineIndent": newLineIndent,
-}
-
 func (c *TemplateConfigurator[T]) Configure(cfgTmpl types.ActionConfig, overwrites types.ActionConfigOverwrites) clierror.Error {
 	tmplBytes, err := yaml.Marshal(cfgTmpl)
 	if err != nil {
 		return clierror.Wrap(err, clierror.New("failed to marshal config template"))
 	}
 
-	configTmpl, err := template.
-		New("config").
-		Option("missingkey=zero").
-		Delims("${{", "}}").Funcs(funcMap).
-		Parse(string(tmplBytes))
-	if err != nil {
-		return clierror.Wrap(err, clierror.New("failed to parse config template"))
+	configBytes, clierr := templateConfig(tmplBytes, overwrites)
+	if clierr != nil {
+		return clierror.WrapE(clierr, clierror.New("failed to template config"))
 	}
 
-	templatedConfig := bytes.NewBuffer([]byte{})
-	err = configTmpl.
-		Execute(templatedConfig, overwrites)
-	if err != nil {
-		return clierror.Wrap(err, clierror.New("failed to template config"))
-	}
-
-	err = yaml.Unmarshal(templatedConfig.Bytes(), &c.Cfg)
+	err = yaml.Unmarshal(configBytes, &c.Cfg)
 	if err != nil {
 		return clierror.Wrap(err, clierror.New("failed to unmarshal config"))
 	}
 
 	return nil
-}
-
-func newLineIndent(n int, s string) string {
-	return strings.ReplaceAll(s, "\n", fmt.Sprintf("\n%s", strings.Repeat(" ", n)))
 }
